requirements.txt
fastapi
uvicorn
redis
requests
python-dotenv
discord.py
